% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{add_months}
\alias{add_months}
\title{Add k months to first-of-month dates (vectorised over both arguments)}
\usage{
add_months(m, k)
}
\description{
Add k months to first-of-month dates (vectorised over both arguments)
}
\keyword{internal}
