% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{code_to_month}
\alias{code_to_month}
\title{Contract month from a YYYYMM contract code}
\usage{
code_to_month(contract_code)
}
\description{
Contract month from a YYYYMM contract code
}
\keyword{internal}
