% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{contract_label}
\alias{contract_label}
\title{Human-readable contract label, locale-independent ("December 2026")}
\usage{
contract_label(contract_month)
}
\description{
Human-readable contract label, locale-independent ("December 2026")
}
\keyword{internal}
