% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{front_contract_month}
\alias{front_contract_month}
\title{Front contract month on each date}
\usage{
front_contract_month(dates, table_code)
}
\arguments{
\item{dates}{Date vector}

\item{table_code}{product table code}
}
\value{
first-of-month Date vector
}
\description{
Front contract month on each date
}
\keyword{internal}
