% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{get_dim_id}
\alias{get_dim_id}
\title{Get the tab_dim_id of a named non-time dimension}
\usage{
get_dim_id(tbl_id, dimension, con, schema = "platform")
}
\description{
Get the tab_dim_id of a named non-time dimension
}
\keyword{internal}
