% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{get_ice_source_id}
\alias{get_ice_source_id}
\title{Get the ICE source id, failing loudly if ICE is not registered}
\usage{
get_ice_source_id(con, schema = "platform")
}
\description{
Get the ICE source id, failing loudly if ICE is not registered
}
\keyword{internal}
