% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{get_ice_table_id}
\alias{get_ice_table_id}
\title{Get a table id, failing loudly if the table is not registered}
\usage{
get_ice_table_id(table_code, con, schema = "platform")
}
\description{
Get a table id, failing loudly if the table is not registered
}
\keyword{internal}
