% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{ice_expiry}
\alias{ice_expiry}
\title{Rule-based last trading day of an ICE contract}
\usage{
ice_expiry(table_code, contract_month)
}
\arguments{
\item{table_code}{product table code}

\item{contract_month}{first-of-month Date(s)}
}
\value{
Date(s)
}
\description{
Rule-based last trading day of an ICE contract
}
