% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{ice_get_history}
\alias{ice_get_history}
\title{Get daily price history for one ICE contract}
\usage{
ice_get_history(market_id, span = 3)
}
\arguments{
\item{market_id}{ICE marketId from `ice_get_contracts()`}

\item{span}{ICE historicalSpan parameter (3 = ~2 years)}
}
\value{
data frame: period_id (YYYY-MM-DD), value
}
\description{
Get daily price history for one ICE contract
}
