% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare row for the `category` table (single root category per source)}
\usage{
prepare_category_table(con, schema = "platform")
}
\arguments{
\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
data frame with id, name, source_id
}
\description{
Prepare row for the `category` table (single root category per source)
}
