% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare row for the `category_table` table}
\usage{
prepare_category_table_table(table_code, con, schema = "platform")
}
\arguments{
\item{table_code}{product table code, e.g. "GASOIL_RAW"}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
data frame with category_id, table_id, source_id
}
\description{
Prepare row for the `category_table` table
}
