% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_source_table}
\alias{prepare_source_table}
\title{Prepare row for the `source` table}
\usage{
prepare_source_table(con, schema = "platform")
}
\arguments{
\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
data frame with id, name, name_long, url, or NULL if ICE already exists
}
\description{
Prepare row for the `source` table
}
