% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_unit_table}
\alias{prepare_unit_table}
\title{Prepare row for the `unit` table}
\usage{
prepare_unit_table(table_code)
}
\arguments{
\item{table_code}{product table code, e.g. "GASOIL_RAW"}
}
\value{
data frame with name
}
\description{
Prepare row for the `unit` table
}
