% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{prev_business_day}
\alias{prev_business_day}
\title{Previous business day (weekends only, no holidays)}
\usage{
prev_business_day(x)
}
\description{
Previous business day (weekends only, no holidays)
}
\keyword{internal}
