% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read_raw_history}
\alias{read_raw_history}
\title{Read the latest stored history of all raw contract series}
\usage{
read_raw_history(table_code, con, schema = "platform")
}
\arguments{
\item{table_code}{raw product table code}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
data frame: contract_code, period_id, value
}
\description{
Read the latest stored history of all raw contract series
}
\keyword{internal}
